fixes from code review
add warning for interfaces

escape hatches

refactored jest tests
