+++ lib/deployment.ts
@@ -35,6 +35,7 @@ export class Deployment extends Construct {
     const volumes: Volume[] | undefined = props.secretMounts?.map(m => new Volume(m))
     const autoScalingOn : boolean = props.autoScalingProps !== undefined
 
+    // See https://github.com/kubernetes/kubernetes/issues/25238#issuecomment-570257435 for info
     if (autoScalingOn && props.replicas !== undefined) {
       throw new Error('Cannot specify "replicaCount" when auto-scaling is enabled');
     }
