+++ test/chart.test.ts
@@ -1,0 +1,13 @@
+import { Construct } from 'constructs';
+import { Application, synth } from '../lib';
+
+export function buildWebsiteChart(scope: Construct) {
+
+  new Application(scope, 'serve', {
+    image: 'pennlabs/website',
+    ingress: [{ host: 'pennlabs.org', paths: ['/'] }],
+  })
+}
+
+test('Synth Function', () =>
+  expect(function () { synth(buildWebsiteChart) }).not.toThrow());
